chore: move TabGroup to Footer
